PracticaFinal/src/backend/requirements.txt
flask==2.3.3
gunicorn==21.2.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-instrumentation-flask==0.40b0
requests==2.31.0